Create README.md
Create an array of objects, where each object represents a quote. Each object should have properties like quote and author.

Write JavaScript code to generate a random number between 0 and the length of the array minus one. This number will be used as the index of the array to retrieve a random quote.

Retrieve the random quote from the array using the index.

Update the quote element on the page with the text content of the retrieved quote.

Add an event listener to the button element to generate a new random quote when the button is clicked.

Optionally, you can add CSS styling to make the layout responsive and visually appealing.